bench: extract common helpers for benchmarks
